feat: tratamento de erros para usuário duplicado e não encontrado

+++ src/user/user.service.ts
@@ -1,7 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, ConflictException, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User } from './user.entity';
+import * as bcrypt from 'bcryptjs';
 
 @Injectable()
 export class UserService {
@@ -11,7 +12,17 @@ export class UserService {
   ) {}
 
   async create(user: Partial<User>): Promise<User> {
-    return this.userRepository.save(user);
+    if (user.password) {
+      user.password = await bcrypt.hash(user.password, 10);
+    }
+    try {
+      return await this.userRepository.save(user);
+    } catch (error) {
+      if (error.code === '23505') { // Postgres unique violation
+        throw new ConflictException('E-mail já cadastrado');
+      }
+      throw error;
+    }
   }
 
   async findAll(): Promise<User[]> {
@@ -19,7 +30,11 @@ export class UserService {
   }
 
   async findOne(id: number): Promise<User | null> {
-    return this.userRepository.findOneBy({ id });
+    const user = await this.userRepository.findOneBy({ id });
+    if (!user) {
+      throw new NotFoundException('Usuário não encontrado');
+    }
+    return user;
   }
 
   async findByEmail(email: string): Promise<User | null> {
@@ -27,11 +42,25 @@ export class UserService {
   }
 
   async update(id: number, user: Partial<User>): Promise<User | null> {
-    await this.userRepository.update(id, user);
-    return this.findOne(id);
+    if (user.password) {
+      user.password = await bcrypt.hash(user.password, 10);
+    }
+    try {
+      await this.userRepository.update(id, user);
+      return this.findOne(id);
+    } catch (error) {
+      if (error.code === '23505') {
+        throw new ConflictException('E-mail já cadastrado');
+      }
+      throw error;
+    }
   }
 
   async remove(id: number): Promise<void> {
+    const user = await this.userRepository.findOneBy({ id });
+    if (!user) {
+      throw new NotFoundException('Usuário não encontrado');
+    }
     await this.userRepository.delete(id);
   }
 }
